organized file structure + edited exercise list
