--- SwiftJym/ExercisesView.swift
@@ -1,14 +1,0 @@
-//
-//  ExercisesView.swift
-//  SwiftJym
-//
-//  Created by Andrea Rota on 12/7/23.
-//
-
-import SwiftUI
-
-struct ExercisesView: View {
-    var body: some View {
-        Text("TODO ExercisesView")
-    }
-}
