+++ SwiftJym/Models/Exercise.swift
